Refactor general post type functions into separate php partial
