+++ functions.php
@@ -11,6 +11,7 @@
 $baseUrl = site_url();
 
 // Required files
+require_once 'functions/post-types/general.php';
 require_once 'functions/theme/setup.php';
 require_once 'functions/theme/minify.php';
 require_once 'functions/customizer/customize-register.php';
@@ -34,6 +35,11 @@ require_once 'functions/theme/custom-fields/image_id.php';
 require_once 'functions/theme/custom-fields/page_links_to.php';
 require_once 'functions/theme/custom-fields/acf-search.php';
 require_once 'functions/tinymce/editor.php';
+
+// Initialize post types if ACF plugin is present and activated
+if ( class_exists('Acf') ) {
+	require_once locate_template('/functions/post-types/init.php');
+}
 
 // Add wide class to alignment options
 add_theme_support ('align-wide');
@@ -471,112 +477,6 @@ function cwd_news_cards_markup($query, $fallback_override = null) {
 
 	include('templates/news/cards.php');
 }
-
-
-
-///////////////////////////////////////////////////////////\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
-///////////////////////////////////////////////// All post type stuff \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
-///////////////////////////////////////////////////////////\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
-
-// Flush rewrite rules when a custom post type post is saved, but only once for each custom post type
-if ( ! function_exists ( 'cwd_flush_rewrite_rules' ) ) {
-	function cwd_flush_rewrite_rules() {
-
-		global $post; 
-		$post_type = $post->post_type;
-						
-		if ( $post_type == 'page' || $post_type == 'post' || $post_type.'_already_flushed' == '1' ) {
-			return;
-		}
-
-		flush_rewrite_rules();
-		
-		add_option($post_type.'_already_flushed', '1', '', 'yes');
-
-	}
-	add_action('save_post', 'cwd_flush_rewrite_rules', 999);
-}
-
-// Get all custom post types
-if ( ! function_exists ( 'get_all_custom_post_types' ) ) {
-	function get_all_custom_post_types() {
-
-		$cptui_post_types = cptui_get_post_type_slugs();
-		$local_post_types = array('courses', 'events', 'galleries', 'news', 'people', 'spotlights');
-		$all_custom_post_types = array_merge($cptui_post_types, $local_post_types);
-
-		return $all_custom_post_types;
-
-	}
-}
-
-if ( ! function_exists ( 'get_all_post_types' ) ) {
-	function get_all_post_types() {
-	
-		$all_custom_post_types = get_all_custom_post_types();
-		$built_in_post_types = array('post', 'page');
-		$all_post_types = array_merge($all_custom_post_types, $built_in_post_types);
-		
-		return $all_post_types;
-		
-	}
-}
-
-// Initialize post types if ACF plugin is present and activated (which it is by now, I guess :))
-if( class_exists('Acf') ) {
-	require_once locate_template('/functions/post-types/init.php');
-}
-
-// Include all post types in all archives
-if ( ! function_exists ( 'cwd_base_cpt_archives' ) ) {
-	function cwd_base_cpt_archives( $query ) {
-		if ( $query->is_tag() || is_category() && $query->is_main_query() && !is_admin() ) {
-			$query->set( 'post_type', get_all_custom_post_types() );
-		}
-	}
-	add_action( 'pre_get_posts', 'cwd_base_cpt_archives' );
-}
-
-// Filter post type metadata choices 
-if ( ! function_exists ( 'cwd_base_cpt_archives' ) ) {
-	function add_metadata_filter() {
-	
-		$checked_post_types = get_checked_post_types();
-
-		foreach($checked_post_types as $post_type) {
-			add_filter('acf/load_field/name=metadata_' . $post_type, 'get_taxonomies_from_post_type');
-		}
-		
-	}
-	add_action('cptui_post_register_taxonomies', 'add_metadata_filter');
-}
-
-// Toggle metatag options for each post type
-/* function metatags_field_visbility ($field) {
-	
-	$post_type_options = get_field('post_type_options', 'options', true);
-	$post_type_choices = $post_type_options['post_types'];
-
-	if (in_array('news', $post_type_choices)) {     
-		return $field; 
-	}
-}
- *///add_filter('acf/load_field/key=field_60e83df931bd2', 'metatags_field_visbility');
-
-/* foreach(get_chosen_post_types() as $chosen_post_type) {
-
-	add_filter('acf/load_field/key=field_6345b945b9d85', function ($field) use ($chosen_post_type) {
-		$field['sub_fields'] = array(
-			array(
-				'label' => ucwords(str_replace('_', ' ', $chosen_post_type)),
-				'name' => $chosen_post_type,
-				'type' => 'group',
-			),
-		);
-		return $field;
-	});
-
-} */
 
 /*
  * Default (test) WP Saml Auth config (see https://wordpress.org/plugins/wp-saml-auth)
